Return Java/RMLMapper error without splicing by default
There can be more ways than we expect that RMLMapper exceptions appear,
so there is no sense in partitioning/splitting the error string in the
default case. For e.g., the following error would not have been caught
otherwise:

```
Error: Unable to access jarfile /app/.rmlmapper/rmlmapper.jar
```

It is a Java error, not an exception from the JAR being executed
(RMLMapper in this case). It does not match any of the existing
conditions so the spliced entry being accessed would not exist.