Add code styling for XPaths, SPARQL, SHACL
It is desireable to view code text formatted nicely, for which we can
exploit the Prism syntax highlighter.

- The SHACL detailed report is simply a table of RDF resources, so all
columns are styled as Turtle.

- It is possible to add a scrollbar in some places simply by not
wrapping (e.g. in the Conceptual Mapping Rules)

- Wrapping needs to be either an overflowWrap or a wordBreak depending
on where the component is placed (table/grid/PropertyListItem).

Otherwise, this needs refactoring in some way to inherit the theme
(there is theme, i.e. "style" support in the highlighter)